test/requirements.txt
retracesoftware.utils
retracesoftware.functional
retracesoftware.stream
opentelemetry-api
opentelemetry-sdk
psycopg2-binary